Add new containerd event tables
Add one containerd_<event>_events table for each of the containerd
event types, which are: container, content, image, namespace,
snapshot and task.
The path to the containerd unix socket can be configured
with the flag containerd_socket.
By default the tables listen on events coming from all namespaces;
to filter them once can set the namespaces with a comma separated list
in the containerd_events_namespaces flag.

Add gRPC library and dependencies to communicate
with the containerd socket.

Co-Authored-By: Akshay K <iit.akshay@gmail.com>