Add new lxd event tables
Add one lxd_<event_type>_events table for each
of the lxd event types, which are: operation, lifecycle,
logging.
The path to the lxd unix socket can be configured
with the flag lxd_socket.
By default the tables listen on all event types;
to filter them one can set the event types with a comma separated list
in the lxd_event_types flag.