squash: Update lzma README and config folder #2s
Update to remove information which is now aggregated under
libraries/cmake/README.md

Update the order of the config folders so that it's os first,
architecture after, given that if the library config has something
in common, it will be across architectures in the same os.

Regenerate the configurations on Big Sur instead of Catalina.